moved compose function into a helpers module
