+++ test/day6.test.js
@@ -5,7 +5,7 @@ var chai = require('chai')
 
 var s = require('../day6/solution.js')
 
-describe('Part One', function(){
+describe('Day Six', function(){
   var testInput = 'eedadn\ndrvtee\neandsr\nraavrd\natevrs\ntsrnev\nsdttsa\nrasrtv\nnssdts\nntnada\nsvetve\ntesnvt\nvntsnd\nvrdear\ndvrsen\nenarar'
   describe('#solvePartOne()', function(){
     it('should solve the test input', function(){
@@ -14,9 +14,8 @@ describe('Part One', function(){
     });
   });
   describe('#solvePartTwo()', function(){
-    it('should solve the test input given the new methodology', function(argument) {
+    it('should solve the test input given the new methodology', function() {
       var res = s.solvePartTwo(testInput)
-      console.log(res)
       expect(res).to.equal('advent')
     });  
   });
